Source/UnitTest/Startup/Startup.UnitTest/PrinterInstalledConditionTest.cs
using System.Collections.Generic;
using NSubstitute;
using NUnit.Framework;
using pdfforge.PDFCreator.Conversion.Settings;
using pdfforge.PDFCreator.Core.Startup.StartConditions;
using pdfforge.PDFCreator.Core.StartupInterface;
using pdfforge.PDFCreator.UI.ViewModels;
using pdfforge.PDFCreator.UI.ViewModels.Assistants;
using pdfforge.PDFCreator.UI.ViewModels.Translations;

namespace pdfforge.PDFCreator.UnitTest.Startup
{
    [TestFixture]
    public class PrinterInstalledConditionTest
    {
        [Test]
        public void WhenNoRepairRequired_Successful()
        {
            var repairPrinterAssitant = Substitute.For<IRepairPrinterAssistant>();
            repairPrinterAssitant.IsRepairRequired().Returns(false);

            var settingsLoader = Substitute.For<ISettingsLoader>();

            var printerInstalledCondition = new PrinterInstalledCondition(repairPrinterAssitant, settingsLoader, new ApplicationTranslation());

            var result = printerInstalledCondition.Check();

            Assert.IsTrue(result.IsSuccessful);
        }

        [Test]
        public void WhenNoRepairFails_CheckFails()
        {
            var repairPrinterAssitant = Substitute.For<IRepairPrinterAssistant>();
            repairPrinterAssitant.IsRepairRequired().Returns(true);
            repairPrinterAssitant.TryRepairPrinter(Arg.Any<IEnumerable<string>>()).Returns(false);

            var settingsLoader = Substitute.For<ISettingsLoader>();
            settingsLoader.LoadPdfCreatorSettings().Returns(new PdfCreatorSettings(null));


            var printerInstalledCondition = new PrinterInstalledCondition(repairPrinterAssitant, settingsLoader, new ApplicationTranslation());

            var result = printerInstalledCondition.Check();

            Assert.IsFalse(result.IsSuccessful);
            Assert.AreEqual((int)ExitCode.PrintersBroken, result.ExitCode);
        }

        [Test]
        public void WhenNoRepairSucceeds_Successful()
        {
            var repairPrinterAssitant = Substitute.For<IRepairPrinterAssistant>();
            repairPrinterAssitant.IsRepairRequired().Returns(true);
            repairPrinterAssitant.TryRepairPrinter(Arg.Any<IEnumerable<string>>()).Returns(true);

            var settingsLoader = Substitute.For<ISettingsLoader>();
            settingsLoader.LoadPdfCreatorSettings().Returns(new PdfCreatorSettings(null));


            var printerInstalledCondition = new PrinterInstalledCondition(repairPrinterAssitant, settingsLoader, new ApplicationTranslation());

            var result = printerInstalledCondition.Check();

            Assert.IsTrue(result.IsSuccessful);
        }

        [Test]
        public void DuringRepair_AllMappedPrintersArePassedToRepairAssistant()
        {
            var settings = new PdfCreatorSettings(null);
            settings.ApplicationSettings.PrinterMappings.Add(new PrinterMapping("Printer1", ""));
            settings.ApplicationSettings.PrinterMappings.Add(new PrinterMapping("Printer2", ""));

            var settingsLoader = Substitute.For<ISettingsLoader>();
            settingsLoader.LoadPdfCreatorSettings().Returns(settings);

            IEnumerable<string> printersToRepair = null;

            var repairPrinterAssitant = Substitute.For<IRepairPrinterAssistant>();
            repairPrinterAssitant.IsRepairRequired().Returns(true);
            repairPrinterAssitant.TryRepairPrinter(Arg.Any<IEnumerable<string>>()).Returns(x =>
            {
                printersToRepair = x.ArgAt<IEnumerable<string>>(0);
                return true;
            });

            var printerInstalledCondition = new PrinterInstalledCondition(repairPrinterAssitant, settingsLoader, new ApplicationTranslation());

            printerInstalledCondition.Check();

            CollectionAssert.AreEquivalent(new[] {"Printer1", "Printer2" }, printersToRepair);
        }
    }
}
